% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_batches.R
\name{check_batch}
\alias{check_batch}
\title{Fish batch checker (AFB)}
\usage{
check_batch(
  id = NULL,
  type = NULL,
  min_size = NULL,
  max_size = NULL,
  nb = NULL,
  ind_measure = NULL,
  ind_id = NULL,
  ind_size = NULL,
  ...
)
}
\arguments{
\item{id}{int id of the batch}

\item{type}{character type of the batch (N, G, S/L, I)}

\item{min_size}{dbl minimum size of the batch}

\item{max_size}{dbl maximum size of the batch}

\item{nb}{int effectif of the batch}

\item{ind_measure}{data.frame individual measurement of the fish}

\item{ind_size}{variable name for the size in ind_measure}
}
\value{
a list. Each element of the list correspond to a batch of a fishing
operation. If there is no problem in the batch, it returns NA. In the opposite
case, the function returns a character containing the type of error.
}
\description{
Check the quality of the batch
}
\details{
The function check if there are missing datas that can make the batch
invalid. The check are specific of each batch type. For batch "G", we check if
min or max size are missing, if min >= max size and the number of fish in the
batch is superior or equal to 10. For batch "S/L", we check if the number of measured
fishes is superior to 20 (standard protocol propose 30 be measured). In batch
"I", we check that the number of measured fishes correspond to the number of
fish recorded in the batch. For batch "N", we check if the number of measured
fish is equal to one.
}
