% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_operation.R
\name{clean_fish_batch}
\alias{clean_fish_batch}
\title{Clean and standardize batch reference data from ASPE database}
\usage{
clean_fish_batch(
  fish_batch = get_fish_batch_aspe(),
  batch_ref = cleaning_ref_batch_type_aspe(),
  species_ref = cleaning_species_ref_aspe(),
  detail_sampling = cleaning_elementary_sampling()
)
}
\arguments{
\item{fish_batch}{A data frame containing raw batch data.
By default uses \code{get_fish_batch_aspe()} to retrieve raw data.}

\item{batch_ref}{A data frame containing cleaned batch type reference data.
By default uses \code{cleaning_ref_batch_type_aspe()} to retrieve cleaned data.}

\item{species_ref}{A data frame containing cleaned species reference data.
By default uses \code{cleaning_species_ref_aspe()} to retrieve cleaned data.}

\item{detail_sampling}{A data frame containing cleaned elementary sampling data.
By default uses \code{cleaning_elementary_sampling()} to retrieve cleaned data.}
}
\value{
A data frame
}
\description{
Clean batch data by renaming columns, and translating them, joining batch names, species code and opration ids.
}
