% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_operation.R
\name{cleaning_ref_batch_type_aspe}
\alias{cleaning_ref_batch_type_aspe}
\title{Clean and standardize batch reference data from ASPE database}
\usage{
cleaning_ref_batch_type_aspe(ref_batch = get_ref_type_batch_aspe())
}
\arguments{
\item{ref_batch}{A data frame containing species reference data.
By default uses \code{get_ref_type_batch_aspe()} to retrieve raw data.}
}
\value{
A data frame
}
\description{
Clean batch reference data by renaming columns, and translating them.
}
