% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_operation.R
\name{cleaning_ref_length_type_aspe}
\alias{cleaning_ref_length_type_aspe}
\title{Clean and standardize length reference data from ASPE database}
\usage{
cleaning_ref_length_type_aspe(ref_length = get_ref_type_length_aspe())
}
\arguments{
\item{ref_length}{A data frame containing length reference data.
By default uses \code{get_ref_type_length_aspe()} to retrieve raw data.}
}
\value{
A data frame
}
\description{
Clean length reference data by renaming columns, and translating them.
}
