% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aspe_data.R
\name{extract_optional_data}
\alias{extract_optional_data}
\title{Extract optional data for the package}
\usage{
extract_optional_data(packagename = "fishdatabuilder")
}
\arguments{
\item{packagename}{string. The name of the package using the package cache. Must be a valid directory name. Should not contain spaces. Passed as 'appname' to `rappdirs::user_data_dir`.}
}
\value{
logical vector indicating if the files have been successfully moved
in the main folder of the package cache
}
\description{
Extract
package cache. Will try to download the data only if it is not available.
}
