% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishdatabuilder-package.R
\docType{package}
\name{fishdatabuilder-package}
\alias{fishdatabuilder}
\alias{fishdatabuilder-package}
\title{fishdatabuilder: Tools to Make Analysis of Fish Monitoring Data Easier}
\description{
Collection of tools to clean, filter, rearrange, and augment fish monitoring data.
}
\author{
\strong{Maintainer}: Alain Danet \email{alain.danet@inrae.fr} (\href{https://orcid.org/0000-0002-1592-9483}{ORCID})

Authors:
\itemize{
  \item Alain Danet \email{alain.danet@inrae.fr} (\href{https://orcid.org/0000-0002-1592-9483}{ORCID})
  \item Camille Leclerc (\href{https://orcid.org/0000-0001-5830-1787}{ORCID})
  \item Willem Bonnaffé (\href{https://orcid.org/0000-0002-5053-8891}{ORCID})
}

}
\keyword{internal}
