% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_batches.R
\name{gen_fish_from_batch}
\alias{gen_fish_from_batch}
\title{Generate fish from fishing batch (AFB)}
\usage{
gen_fish_from_batch(
  id = NULL,
  type = NULL,
  min_size = NULL,
  max_size = NULL,
  nb = NULL,
  ind_measure = NULL,
  id_var = NULL,
  ind_size = NULL,
  ...
)
}
\arguments{
\item{id}{int id of the batch}

\item{type}{character type of the batch (N, G, S/L, I)}

\item{min_size}{dbl minimum size of the batch}

\item{max_size}{dbl maximum size of the batch}

\item{nb}{int effectif of the batch}

\item{ind_measure}{data.frame individual measurement of the fish}

\item{ind_size}{variable name for the size in ind_measure}
}
\description{
Generate fish from fishing batch (AFB)
}
\details{
From fishing batch, we build generate the fish size individual
}
