% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_batches.R
\name{get_check_batch}
\alias{get_check_batch}
\title{Main function for batch check}
\usage{
get_check_batch(
  batch = NULL,
  id_var = NULL,
  batch_type_var = NULL,
  nb_var = NULL,
  min_var = NULL,
  max_var = NULL,
  species = NULL,
  measure = NULL,
  size_var = NULL,
  future_enabled = FALSE,
  ...
)
}
\description{
return a data.frame with batch id, species and the
}
