% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aspe_data.R
\name{get_optional_filepath}
\alias{get_optional_filepath}
\title{Access a single file from the package cache by its file name.}
\usage{
get_optional_filepath(
  packagename = "fishdatabuilder",
  filename = NULL,
  mustWork = TRUE
)
}
\arguments{
\item{packagename}{string. The name of the package using the package cache. Must be a valid directory name. Should not contain spaces. Passed as 'appname' to `rappdirs::user_data_dir`.}

\item{filename, }{string. A filename, relative to the package cache. Passed to  \code{relative_filename} argument in pkgfilecache::get_filepath().}

\item{mustWork}{logical. Whether an error should be created if the file does not exist.}
}
\value{
string. The full path to the file in the package cache. Use this in
your application code to open the file.
}
\description{
Access a single file from the package cache by its file name.
}
