% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_batches.R
\name{get_size_from_batch}
\alias{get_size_from_batch}
\title{size from batch (AFB)}
\usage{
get_size_from_batch(
  batch = NULL,
  id_var = NULL,
  batch_type_var = NULL,
  nb_var = NULL,
  min_var = NULL,
  max_var = NULL,
  species = NULL,
  measure = NULL,
  size_var = NULL,
  future_enabled = FALSE,
  ...
)
}
\arguments{
\item{batch}{data.frame}
}
\description{
size from batch (AFB)
}
