% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aspe_data.R
\name{list_optional_files}
\alias{list_optional_files}
\title{Get file names available in package cache}
\usage{
list_optional_files(packagename = "fishdatabuilder")
}
\arguments{
\item{packagename}{string. The name of the package using the package cache. Must be a valid directory name. Should not contain spaces. Passed as 'appname' to `rappdirs::user_data_dir`.}
}
\value{
vector of strings. The file names available, relative to the package
cache.
}
\description{
Get file names of optional data files which are available in the
local package cache. You can access these files with
optional_filepath().
}
