% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_species.R
\name{species_latin_name_to_replace}
\alias{species_latin_name_to_replace}
\title{Replacement vector to fix species latin name}
\usage{
species_latin_name_to_replace()
}
\value{
named vector
}
\description{
Replacement vector to fix species latin name
}
