% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_operation.R
\name{clean_individual_measurement_aspe}
\alias{clean_individual_measurement_aspe}
\title{Clean individual measurement data from ASPE database}
\usage{
clean_individual_measurement_aspe(
  ind_measure = get_individual_measurement_aspe()
)
}
\arguments{
\item{ind_measure}{A data frame containing the raw data.
By default uses \code{get_individual_measurement_aspe()} to retrieve raw data.}
}
\value{
A data frame
}
\description{
Clean data by renaming columns, and translating them. It checks also that all measurements are real measurements.
}
