% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_operation.R
\name{cleaning_species_ref_aspe}
\alias{cleaning_species_ref_aspe}
\title{Clean and standardize species reference data from ASPE database}
\usage{
cleaning_species_ref_aspe(species = get_species_aspe())
}
\arguments{
\item{species}{A data frame containing species reference data.
By default uses \code{get_species_aspe()} to retrieve raw data.}
}
\value{
A data frame
}
\description{
Clean species reference data by renaming species columns, and translating them.
}
