% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aspe_data.R
\name{extract_optional_data}
\alias{extract_optional_data}
\title{Extract optional data for the package}
\usage{
extract_optional_data()
}
\value{
logical vector indicating if the files have been successfully moved
in the main folder of the package cache
}
\description{
Extract
package cache. Will try to download the data only if it is not available.
}
