% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_strings.R
\name{replacement_station_col}
\alias{replacement_station_col}
\title{Replacement vector to replace station.csv column names}
\usage{
replacement_station_col()
}
\value{
named vector
}
\description{
Replacement vector to replace station.csv column names
}
