% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_fish_size.R
\name{sanatize_size_aspe}
\alias{sanatize_size_aspe}
\title{Sanitize fish size measurements (deprecated)}
\usage{
sanatize_size_aspe(
  fish_batch = clean_fish_batch(),
  ind_measure = clean_individual_measurement_aspe(),
  species_ref = cleaning_species_ref_aspe(),
  fishbase_length_length = NULL,
  conversion_vector = coefficients_fork2total(),
  convert_intercept_cm2mm = TRUE
)
}
\arguments{
\item{...}{Arguments passed to the new functions}
}
\description{
This function is deprecated. Please use the more focused functions:
\code{\link{convert_fork_to_total}} and \code{\link{remove_impossible_lengths}}.
}
